+++ runtime/syscall-server/CMakeLists.txt
@@ -5,7 +5,10 @@ project(
 )
 
 add_library(bpftime-syscall-server SHARED
-    syscall_server.cpp)
+    syscall_context.cpp
+    syscall_server_main.cpp
+    syscall_server_utils.cpp
+)
 target_link_libraries(bpftime-syscall-server PUBLIC
     runtime
     -lpthread
