+++ runtime/syscall-server/syscall_context.cpp
@@ -1,226 +1,11 @@
-#include "syscall_server.hpp"
+#include "syscall_context.hpp"
 #include "bpftime_shm.hpp"
 #include "handler/perf_event_handler.hpp"
-#include "linux/bpf.h"
-#include "spdlog/cfg/env.h"
-#include <asm-generic/errno-base.h>
-#include <cstdlib>
-#include <cstring>
+#include <linux/bpf.h>
+#include "syscall_server_utils.hpp"
 #include <optional>
-#include <ranges>
-#include <spdlog/spdlog.h>
-#include <string_view>
-#include <unistd.h>
-#ifdef ENABLE_BPFTIME_VERIFIER
-#include <bpftime-verifier.hpp>
-#include <iomanip>
-#include <sstream>
-#endif
-
 #include <sys/epoll.h>
-#include <spdlog/cfg/env.h>
 using namespace bpftime;
-
-const shm_open_type bpftime::global_shm_open_type = shm_open_type::SHM_SERVER;
-
-// global context for bpf syscall server
-static syscall_context context;
-
-#define PERF_UPROBE_REF_CTR_OFFSET_BITS 32
-#define PERF_UPROBE_REF_CTR_OFFSET_SHIFT 32
-
-static bool already_setup = false;
-
-static void start_up()
-{
-	if (already_setup)
-		return;
-	spdlog::cfg::load_env_levels();
-	spdlog::set_pattern("[%Y-%m-%d %H:%M:%S][%^%l%$][%t] %v");
-	auto &agent_config = bpftime_get_agent_config();
-	if (const char *custom_helpers = getenv("BPFTIME_HELPER_GROUPS");
-	    custom_helpers != nullptr) {
-		agent_config.enable_kernel_helper_group =
-			agent_config.enable_ffi_helper_group =
-				agent_config.enable_shm_maps_helper_group =
-					false;
-		auto helpers_sv = std::string_view(custom_helpers);
-		for (auto tok : helpers_sv | std::ranges::views::split(
-						     std::string_view(","))) {
-			auto curr_token =
-				std::string_view(tok.begin(), tok.end());
-			if (curr_token == "ffi") {
-				spdlog::info("Enabling ffi helper group");
-				agent_config.enable_ffi_helper_group = true;
-			} else if (curr_token == "kernel") {
-				spdlog::info("Enabling kernel helper group");
-				agent_config.enable_kernel_helper_group = true;
-			} else if (curr_token == "shm_map") {
-				spdlog::info("Enabling shm_map helper group");
-				agent_config.enable_shm_maps_helper_group =
-					true;
-			} else {
-				spdlog::warn("Unknown helper group: {}",
-					     curr_token);
-			}
-		}
-	} else {
-		spdlog::info(
-			"Enabling helper groups ffi, kernel, shm_map by default");
-		agent_config.enable_kernel_helper_group =
-			agent_config.enable_shm_maps_helper_group =
-				agent_config.enable_ffi_helper_group = true;
-	}
-#ifdef ENABLE_BPFTIME_VERIFIER
-	std::vector<int32_t> helper_ids;
-	std::map<int32_t, bpftime::verifier::BpftimeHelperProrotype>
-		non_kernel_helpers;
-	if (agent_config.enable_kernel_helper_group) {
-		for (auto x :
-		     bpftime_helper_group::get_kernel_utils_helper_group()
-			     .get_helper_ids()) {
-			helper_ids.push_back(x);
-		}
-	}
-	if (agent_config.enable_shm_maps_helper_group) {
-		for (auto x : bpftime_helper_group::get_shm_maps_helper_group()
-				      .get_helper_ids()) {
-			helper_ids.push_back(x);
-		}
-	}
-	if (agent_config.enable_ffi_helper_group) {
-		for (auto x : bpftime_helper_group::get_shm_maps_helper_group()
-				      .get_helper_ids()) {
-			helper_ids.push_back(x);
-		}
-		// non_kernel_helpers =
-		for (const auto &[k, v] : get_ffi_helper_protos()) {
-			non_kernel_helpers[k] = v;
-		}
-	}
-	verifier::set_available_helpers(helper_ids);
-	spdlog::info("Enabling {} helpers", helper_ids.size());
-	verifier::set_non_kernel_helpers(non_kernel_helpers);
-#endif
-	already_setup = true;
-}
-
-/*
- * this function is expected to parse integer in the range of [0, 2^31-1] from
- * given file using scanf format string fmt. If actual parsed value is
- * negative, the result might be indistinguishable from error
- */
-static int parse_uint_from_file(const char *file, const char *fmt)
-{
-	int err, ret;
-	FILE *f;
-
-	f = fopen(file, "re");
-	if (!f) {
-		err = -errno;
-		spdlog::error("Failed to open {}: {}", file, err);
-		return err;
-	}
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wformat-nonliteral"
-	err = fscanf(f, fmt, &ret);
-#pragma GCC diagnostic pop
-	if (err != 1) {
-		err = err == EOF ? -EIO : -errno;
-		spdlog::error("Failed to parse {}: {}", file, err);
-		fclose(f);
-		return err;
-	}
-	fclose(f);
-	return ret;
-}
-
-static int determine_uprobe_perf_type(void)
-{
-	const char *file = "/sys/bus/event_source/devices/uprobe/type";
-
-	return parse_uint_from_file(file, "%d\n");
-}
-
-static int determine_uprobe_retprobe_bit(void)
-{
-	const char *file =
-		"/sys/bus/event_source/devices/uprobe/format/retprobe";
-
-	return parse_uint_from_file(file, "config:%d\n");
-}
-
-extern "C" int epoll_wait(int epfd, epoll_event *evt, int maxevents,
-			  int timeout)
-{
-	spdlog::debug("epoll_wait {}", epfd);
-	return context.handle_epoll_wait(epfd, evt, maxevents, timeout);
-}
-
-extern "C" int epoll_ctl(int epfd, int op, int fd, epoll_event *evt)
-{
-	spdlog::debug("epoll_ctl {} {} {} {}", epfd, op, fd, (uintptr_t)evt);
-	return context.handle_epoll_ctl(epfd, op, fd, evt);
-}
-
-extern "C" int epoll_create1(int flags)
-{
-	spdlog::debug("epoll_create1 {}", flags);
-	return context.handle_epoll_create1(flags);
-}
-
-extern "C" int ioctl(int fd, unsigned long req, int data)
-{
-	spdlog::debug("ioctl {} {} {}", fd, req, data);
-	return context.handle_ioctl(fd, req, data);
-}
-
-extern "C" void *mmap64(void *addr, size_t length, int prot, int flags, int fd,
-			off64_t offset)
-{
-	spdlog::debug("mmap64 {:x}", (uintptr_t)addr);
-	return context.handle_mmap64(addr, length, prot, flags, fd, offset);
-}
-
-extern "C" int close(int fd)
-{
-	spdlog::debug("Closing {}", fd);
-	return context.handle_close(fd);
-}
-
-extern "C" long syscall(long sysno, ...)
-{
-	// glibc directly reads the arguments without considering
-	// the underlying argument number. So did us
-	va_list args;
-	va_start(args, sysno);
-	long arg1 = va_arg(args, long);
-	long arg2 = va_arg(args, long);
-	long arg3 = va_arg(args, long);
-	long arg4 = va_arg(args, long);
-	long arg5 = va_arg(args, long);
-	long arg6 = va_arg(args, long);
-	va_end(args);
-	if (sysno == __NR_bpf) {
-		spdlog::debug("SYS_BPF {} {} {} {} {} {}", arg1, arg2, arg3,
-			      arg4, arg5, arg6);
-		int cmd = (int)arg1;
-		auto attr = (union bpf_attr *)(uintptr_t)arg2;
-		auto size = (size_t)arg3;
-		return context.handle_sysbpf(cmd, attr, size);
-	} else if (sysno == __NR_perf_event_open) {
-		spdlog::debug("SYS_PERF_EVENT_OPEN {} {} {} {} {} {}", arg1,
-			      arg2, arg3, arg4, arg5, arg6);
-		return context.handle_perfevent(
-			(perf_event_attr *)(uintptr_t)arg1, (pid_t)arg2,
-			(int)arg3, (int)arg4, (unsigned long)arg5);
-	} else if (sysno == __NR_ioctl) {
-		spdlog::debug("SYS_IOCTL {} {} {} {} {} {}", arg1, arg2, arg3,
-			      arg4, arg5, arg6);
-	}
-	return context.orig_syscall_fn(sysno, arg1, arg2, arg3, arg4, arg5,
-				       arg6);
-}
 
 int syscall_context::handle_close(int fd)
 {
