+++ runtime/syscall-server/syscall_server_utils.cpp
@@ -1,0 +1,130 @@
+#include <ranges>
+#include <spdlog/cfg/env.h>
+#include <spdlog/spdlog.h>
+#include <bpftime_shm.hpp>
+#ifdef ENABLE_BPFTIME_VERIFIER
+#include <bpftime-verifier.hpp>
+#include <iomanip>
+#include <sstream>
+#endif
+static bool already_setup = false;
+using namespace bpftime;
+
+void start_up()
+{
+	if (already_setup)
+		return;
+	spdlog::cfg::load_env_levels();
+	spdlog::set_pattern("[%Y-%m-%d %H:%M:%S][%^%l%$][%t] %v");
+	auto &agent_config = bpftime_get_agent_config();
+	if (const char *custom_helpers = getenv("BPFTIME_HELPER_GROUPS");
+	    custom_helpers != nullptr) {
+		agent_config.enable_kernel_helper_group =
+			agent_config.enable_ffi_helper_group =
+				agent_config.enable_shm_maps_helper_group =
+					false;
+		auto helpers_sv = std::string_view(custom_helpers);
+		for (auto tok : helpers_sv | std::ranges::views::split(
+						     std::string_view(","))) {
+			auto curr_token =
+				std::string_view(tok.begin(), tok.end());
+			if (curr_token == "ffi") {
+				spdlog::info("Enabling ffi helper group");
+				agent_config.enable_ffi_helper_group = true;
+			} else if (curr_token == "kernel") {
+				spdlog::info("Enabling kernel helper group");
+				agent_config.enable_kernel_helper_group = true;
+			} else if (curr_token == "shm_map") {
+				spdlog::info("Enabling shm_map helper group");
+				agent_config.enable_shm_maps_helper_group =
+					true;
+			} else {
+				spdlog::warn("Unknown helper group: {}",
+					     curr_token);
+			}
+		}
+	} else {
+		spdlog::info(
+			"Enabling helper groups ffi, kernel, shm_map by default");
+		agent_config.enable_kernel_helper_group =
+			agent_config.enable_shm_maps_helper_group =
+				agent_config.enable_ffi_helper_group = true;
+	}
+#ifdef ENABLE_BPFTIME_VERIFIER
+	std::vector<int32_t> helper_ids;
+	std::map<int32_t, bpftime::verifier::BpftimeHelperProrotype>
+		non_kernel_helpers;
+	if (agent_config.enable_kernel_helper_group) {
+		for (auto x :
+		     bpftime_helper_group::get_kernel_utils_helper_group()
+			     .get_helper_ids()) {
+			helper_ids.push_back(x);
+		}
+	}
+	if (agent_config.enable_shm_maps_helper_group) {
+		for (auto x : bpftime_helper_group::get_shm_maps_helper_group()
+				      .get_helper_ids()) {
+			helper_ids.push_back(x);
+		}
+	}
+	if (agent_config.enable_ffi_helper_group) {
+		for (auto x : bpftime_helper_group::get_shm_maps_helper_group()
+				      .get_helper_ids()) {
+			helper_ids.push_back(x);
+		}
+		// non_kernel_helpers =
+		for (const auto &[k, v] : get_ffi_helper_protos()) {
+			non_kernel_helpers[k] = v;
+		}
+	}
+	verifier::set_available_helpers(helper_ids);
+	spdlog::info("Enabling {} helpers", helper_ids.size());
+	verifier::set_non_kernel_helpers(non_kernel_helpers);
+#endif
+	already_setup = true;
+}
+
+/*
+ * this function is expected to parse integer in the range of [0, 2^31-1] from
+ * given file using scanf format string fmt. If actual parsed value is
+ * negative, the result might be indistinguishable from error
+ */
+static int parse_uint_from_file(const char *file, const char *fmt)
+{
+	int err, ret;
+	FILE *f;
+
+	f = fopen(file, "re");
+	if (!f) {
+		err = -errno;
+		spdlog::error("Failed to open {}: {}", file, err);
+		return err;
+	}
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wformat-nonliteral"
+	err = fscanf(f, fmt, &ret);
+#pragma GCC diagnostic pop
+	if (err != 1) {
+		err = err == EOF ? -EIO : -errno;
+		spdlog::error("Failed to parse {}: {}", file, err);
+		fclose(f);
+		return err;
+	}
+	fclose(f);
+	return ret;
+}
+
+int determine_uprobe_perf_type()
+{
+	const char *file = "/sys/bus/event_source/devices/uprobe/type";
+
+	return parse_uint_from_file(file, "%d\n");
+}
+
+int determine_uprobe_retprobe_bit()
+{
+	const char *file =
+		"/sys/bus/event_source/devices/uprobe/format/retprobe";
+
+	return parse_uint_from_file(file, "config:%d\n");
+}
